Fixed full-width images on an own line in post content are wrapped into paragraphs. (#5)
This causes problems to apply a consistent styling for paragraphs and images in CSS, especially concerning the spacing between subsequently following paragraphs with an image in between.

The code unconditionally outputs a figure HTML5 element without checking whether the theme supports HTML5, because that is the assumed default for all themes in 2017.