Fixed Google Analytics opt-out cookie is not set for subdomains by default.
When no COOKIE_DOMAIN is configured, the cookie is set without domain, causing the current document.domain to be used, but without leading period.

A leading period is required for the cookie to be set for all subdomains. Therefore, the fallback/default value is manually set to the current document.domain including leading period.